requirements.txt
asgiref==3.9.2
certifi==2025.10.5
charset-normalizer==3.4.3
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.8.2
